% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_testing.R
\name{get_index}
\alias{get_index}
\title{Performs repeatead rarefaction once and calculates a performance value}
\usage{
get_index(physeq, sample_info, repeats, threshold, method, sample_id, groupb)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{sample_info}{Sample Data object from the phyloseq package. Contains data info without repetitions.}

\item{repeats}{A positive integer. Indicates the amount of repeats run. A value = 1 means no repeats are used.}

\item{threshold}{A positive integer. The threshold value for the rarefaction.}

\item{method}{A string. Currently only "NMDS" supported.}

\item{sample_id}{A string. Name of the column with sample IDs.}

\item{groupb}{A string. Name of the column with data to group the points by.}
}
\value{
Returns the Calinski-Harabasz F-statistic score.
}
\description{
Performs repeatead rarefaction once and calculates a performance value
}
