% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{ord_and_mean}
\alias{ord_and_mean}
\title{Calculates ordination on the repeated count input and as well as a median result for each original datapoint.}
\usage{
ord_and_mean(
  repeat_count,
  repeat_info,
  sample_info,
  repeats,
  method,
  sample_id
)
}
\arguments{
\item{repeat_count}{A matrix. Contains the repeated count data.}

\item{repeat_info}{A matrix. Contains the repeated data info.}

\item{sample_info}{Sample Data object from the phyloseq package. Contains data info without repetitions.}

\item{repeats}{An integer. The amount of repeats that has been performed on the data.}

\item{method}{A string. Currently only "NMDS" supported.}

\item{sample_id}{A string. Name of the column with sample IDs.}
}
\value{
Returns a list containg an ordination object, a phyloseq object with
the repeated count data, a dataframe with all positions from the ordination
calculaction, and a datafram with just the median position from the calculation.
}
\description{
Calculates ordination on the repeated count input and as well as a median result for each original datapoint.
}
