% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{plot_rep_raref}
\alias{plot_rep_raref}
\title{Create a plot for the repeated rarefaction.}
\usage{
plot_rep_raref(
  ordination,
  physeq,
  all_positions,
  median_positions,
  color,
  shape,
  cloud,
  ellipse
)
}
\arguments{
\item{ordination}{An ordination object.}

\item{physeq}{A phyloseq object containing the data to be plotted.}

\item{all_positions}{A dataframe containing position data for all repeats.}

\item{color}{A string. Name of the column to color points by.}

\item{shape}{A string. Name of the column to shape points by.}

\item{cloud}{Boolean. Aesthetic setting for the graph. Default is FALSE.
TRUE shows datapoints for all repeats.
#' @param ellipse Boolean. Aesthetic setting for the graph. Default is TRUE.}

\item{mediant_positions}{A dataframe containing median position data for each original data point.}
}
\value{
Returns an ordination plot.
}
\description{
Create a plot for the repeated rarefaction.
}
