% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{rep_raref}
\alias{rep_raref}
\title{Rarefaction is perforned reaptedly depending on input and a matrix containing all data is created together with a an info file reflecting it.}
\usage{
rep_raref(count, info, threshold, repeats)
}
\arguments{
\item{count}{A matrix. An otu-table containing the count data the rarefaction should be performed on.}

\item{info}{Sample Data object from the phyloseq package.
Should contain info about the count data with matching column names.}

\item{threshold}{An integer. The threshold value at which to perform rarefaction.}

\item{repeats}{An integer. The amount of repeated rarefactions to perform.
A value = 1 means only one iteration of rarefaction is perfomed and therefore no repeats.}
}
\value{
A list containing a matrix with the repeated count table and another
matrix with the repeated info.
}
\description{
Rarefaction is perforned reaptedly depending on input and a matrix containing all data is created together with a an info file reflecting it.
}
