% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{repeated_rarefaction}
\alias{repeated_rarefaction}
\title{Perform one iteration of repeated rarefaction and produce ordination plot}
\usage{
repeated_rarefaction(
  physeq,
  repeats,
  threshold,
  method,
  sample_id,
  colorb,
  shapeb,
  cloud = FALSE,
  ellipse = TRUE
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{repeats}{A positive integer.}

\item{threshold}{A positive integer.}

\item{method}{A string.}

\item{sample_id}{A string.}

\item{colorb}{A string.}

\item{shapeb}{A string.}

\item{cloud}{Boolean}

\item{ellipse}{Boolean}
}
\value{
A list with repeat count table, repeat info table, ordination object, physeq object, dataframe with all repeat ordination positions, dataframe with median ordination positions, and ordination plot.
}
\description{
Perform one iteration of repeated rarefaction and produce ordination plot
}
\examples{
repeated_rarefaction(HLCYG_physeq_data, 5, 500, "NMDS", "sample_id", "location", "type", T, F)
repeated_rarefaction(HLCYG_physeq_data, 10, 250, "NMDS", "sample_id", "location", "type", T, T)
}
