% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_testing.R
\name{test_threshold}
\alias{test_threshold}
\title{Test a range of threshold values for repeated rarefaction.}
\usage{
test_threshold(
  physeq,
  repeats = 10,
  t_min = 5,
  t_max = 250,
  t_step = 1,
  method = "NMDS",
  sample_id,
  groupb
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{repeats}{A positive integer. Indicates the amount of repeats run. A value = 1 means no repeats are used.}

\item{t_min}{An integer. Minimum value for the threshold testing range.}

\item{t_max}{An integer. Maximum value for the threshold testing range.}

\item{t_step}{An float. Step value for the threshold testing range. A value between 0 and 1 will cause the same threshold to be tested more than once.}

\item{method}{A string. Currently only "NMDS" supported.}

\item{sample_id}{A string. Name of the column with sample IDs.}

\item{groupb}{A string. Name of the column with data to group the points by.}
}
\value{
A graph with rarefaction threshold as the x axis, and the y axis is the calculated Calinski-Harabasz F-statistic. Each dot is the score for one threshold attempt. The higher score the better.
}
\description{
Test a range of threshold values for repeated rarefaction.
}
\examples{
test_threshold(HLCYG_physeq_data, repeats = 5, t_step = 10, method= "NMDS", sample_id = "sample_id", groupb = "location")
}
